[decoupling execution] starter task 1
