+++ config/src/config/test_data/validator.yaml
@@ -14,10 +14,10 @@ consensus:
         service:
             type: process
             server_address: "/ip4/127.0.0.1/tcp/5555"
+    decoupled: false
 
 execution:
     genesis_file_location: "relative/path/to/genesis"
-    decoupled: false
 
 full_node_networks:
     - listen_address: "/ip4/0.0.0.0/tcp/7180"
