--- consensus/src/decoupled_state_computer.rs
@@ -1,128 +1,0 @@
-// Copyright (c) The Diem Core Contributors
-// SPDX-License-Identifier: Apache-2.0
-
-use crate::{error::StateSyncError, state_replication::StateComputer};
-use anyhow::Result;
-use consensus_types::{block::Block, executed_block::ExecutedBlock};
-use diem_crypto::{HashValue, hash::ACCUMULATOR_PLACEHOLDER_HASH};
-use diem_infallible::Mutex;
-use diem_logger::prelude::*;
-use diem_metrics::monitor;
-use diem_types::ledger_info::LedgerInfoWithSignatures;
-use execution_correctness::ExecutionCorrectness;
-use executor_types::{Error as ExecutionError, StateComputeResult};
-use fail::fail_point;
-use state_sync::client::StateSyncClient;
-use std::{boxed::Box, sync::Arc};
-
-/// Basic communication with the Execution module;
-/// implements StateComputer traits.
-pub struct DecoupledExecutionProxy {
-    execution_correctness_client: Mutex<Box<dyn ExecutionCorrectness + Send + Sync>>,
-    synchronizer: StateSyncClient,
-}
-
-impl DecoupledExecutionProxy {
-    pub fn new(
-        execution_correctness_client: Box<dyn ExecutionCorrectness + Send + Sync>,
-        synchronizer: StateSyncClient,
-    ) -> Self {
-        Self {
-            execution_correctness_client: Mutex::new(execution_correctness_client),
-            synchronizer,
-        }
-    }
-}
-
-#[async_trait::async_trait]
-impl StateComputer for DecoupledExecutionProxy {
-    fn compute(
-        &self,
-        // The block to be executed.
-        block: &Block,
-        // The parent block id.
-        parent_block_id: HashValue,
-    ) -> Result<StateComputeResult, ExecutionError> {
-        fail_point!("consensus::compute", |_| {
-            Err(ExecutionError::InternalError {
-                error: "Injected error in compute".into(),
-            })
-        });
-        debug!(
-            block_id = block.id(),
-            parent_id = block.parent_id(),
-            "Executing block",
-        );
-
-        // Return dummy block and bypass the execution phase.
-        // This will break the e2e smoke test.
-        monitor!(
-            "execute_block",
-            Ok(StateComputeResult::new(
-                *ACCUMULATOR_PLACEHOLDER_HASH,
-                vec![],
-                0,
-                vec![],
-                0,
-                None,
-                vec![],
-                vec![],
-                vec![],
-            ))
-        )
-
-    }
-
-    /// Send a successful commit. A future is fulfilled when the state is finalized.
-    async fn commit(
-        &self,
-        blocks: &[Arc<ExecutedBlock>],
-        finality_proof: LedgerInfoWithSignatures,
-    ) -> Result<(), ExecutionError> {
-        let mut block_ids = Vec::new();
-        let mut txns = Vec::new();
-        let mut reconfig_events = Vec::new();
-
-        for block in blocks {
-            block_ids.push(block.id());
-            txns.extend(block.transactions_to_commit());
-            reconfig_events.extend(block.compute_result().reconfig_events().to_vec());
-        }
-
-        monitor!(
-            "commit_block",
-            self.execution_correctness_client
-                .lock()
-                .commit_blocks(block_ids, finality_proof)?
-        );
-
-        if let Err(e) = monitor!(
-            "notify_state_sync",
-            self.synchronizer.commit(txns, reconfig_events).await
-        ) {
-            error!(error = ?e, "Failed to notify state synchronizer");
-        }
-        Ok(())
-    }
-
-    /// Synchronize to a commit that not present locally.
-    async fn sync_to(&self, target: LedgerInfoWithSignatures) -> Result<(), StateSyncError> {
-        fail_point!("consensus::sync_to", |_| {
-            Err(anyhow::anyhow!("Injected error in sync_to").into())
-        });
-        // Here to start to do state synchronization where ChunkExecutor inside will
-        // process chunks and commit to Storage. However, after block execution and
-        // commitments, the the sync state of ChunkExecutor may be not up to date so
-        // it is required to reset the cache of ChunkExecutor in State Sync
-        // when requested to sync.
-        let res = monitor!("sync_to", self.synchronizer.sync_to(target).await);
-        // Similarily, after the state synchronization, we have to reset the cache
-        // of BlockExecutor to guarantee the latest committed state is up to date.
-        self.execution_correctness_client.lock().reset()?;
-
-        res.map_err(|error| {
-            let anyhow_error: anyhow::Error = error.into();
-            anyhow_error.into()
-        })
-    }
-}
