+++ consensus/src/lib.rs
@@ -27,7 +27,7 @@ mod pending_votes;
 mod persistent_liveness_storage;
 mod round_manager;
 mod state_computer;
-mod decoupled_state_computer;
+mod ordering_state_computer;
 mod state_replication;
 #[cfg(any(test, feature = "fuzzing"))]
 mod test_utils;
